Split up runtime resource handler into processor manager

PR splits out the runtime authorizer and resource processor into a
dedicated authorizer package, and a manager in the processor package.
This de-couples the main runtime package from the processor so that hot
reloading (not implemented here) can also make use of the processor
manager.

Components are no longer indexed in the ComponentStore by both their name
_and_ type. This is because the hot reloader would not be able to
determine "which" Component it should delete on delete events. No two
Components with the same type but different names can co-exist anyway.
